requirements.txt
asgiref==3.9.1
Django==5.2.6
sqlparse==0.5.3
tzdata==2025.2
djangorestframework==3.15.2
Pillow==10.4.0
django-filter==24.2
